⛵ onion: /repo also available at /fdroid/repo
We want to allow to use our official onion site with the default fdroid
repository path. Since the official repo is hosted at
https://f-droid.org/repo and our onion site is just running a reverse
proxy on top of that deployment the onion site doesn't work as a mirror.
F-Droid client expects mirrors to host the repo under following path:
/fdroid/repo. So this change add /fdroid/repo as an additional location
to the nginx config of our onion site.